CAT08 graph maker - some progresses
Now it works for CAT08/00, but it has problems for all of the others.
Trying to fix it, some more logic was added.
The problem is that the elapsed space along the curves
is such that points with the same linearised distance from its beginning
become very far from the other lines. This means that, for the same "s",
the points along the lines diverge,
making this whole logic unfit.