Redefined nodes.py to hold nodes dictionaries specifications
This is necessary because NetworkX holds all data in dictionary format.
Since this tool will be interproject, I want nodes and edges to be
standardised. Here, I offer a sort of
"blueprint", or better a fixed layout for every kind of node
of our interest.

As of now, just the simple centerline nodes are defined.