Graph save and load created so that datatypes get restored
As of now, there is support only for lists, nested lists, and numpy.ndarrays.
Support for other datatypes can be added (for example torch tensors.
This is suppoorted for attributes of
nodes, edges, and graph.
Note that graph id are automatically
and unavoidably converted to
strings when parsed back from file
to variable.