Added check for None fields in saveGraph
Not saveGraph has fuinctionality to check
that no fields (features) of nodes, edges, or graph
are left empty (None).
If so, it aborts saving and raise error.
This makes unnecessary validating the node HCATNetwork dict
before passing it to NetworkX if the sole purpose of the network is to be
saved to file.